+++ lib/ui/widget/elementary/types.dart
@@ -1,0 +1,3 @@
+import 'package:elementary/elementary.dart';
+
+typedef ListenableEntityState<T> = ListenableState<EntityState<T>>;
